fix: moved delete button
